Release 1.5.2

See merge request ained/charon!477

+++ plugin/app/Repositories/CharonRepository.php
@@ -4,6 +4,7 @@ namespace TTU\Charon\Repositories;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Log;
 use TTU\Charon\Exceptions\CharonNotFoundException;
 use TTU\Charon\Models\Charon;
@@ -145,7 +146,7 @@ class CharonRepository
             throw new CharonNotFoundException('charon_course_module_not_found', $courseModuleId);
         }
 
-        return Charon::with('defenseLabs', 'testerType', 'gradingMethod', 'grademaps.gradeItem', 'deadlines', 'deadlines.group', 'grouping')
+        return Charon::with('defenseLabs', 'testerType', 'gradingMethod', 'grademaps.gradeItem', 'deadlines', 'deadlines.group', 'grouping', 'templates')
             ->where('id', $courseModule->instance)
             ->first();
     }
@@ -161,6 +162,11 @@ class CharonRepository
      */
     public function deleteByInstanceId($id)
     {
+        $id = intval($id);
+        if (intval($id) == 0) {
+            return false;
+        }
+
         /** @var Charon $charon */
         $charon = Charon::find($id);
 
@@ -172,14 +178,19 @@ class CharonRepository
         Deadline::where('charon_id', $id)->delete();
         CharonDefenseLab::where('charon_id', $id)->delete();
 
-        $result = $charon->delete();
+        if ($charon != null) {
 
-        $this->gradebookService->deleteGradeCategory(
-            $charon->category_id,
-            $charon->course
-        );
+            $result = $charon->delete();
 
-        return $result;
+            $this->gradebookService->deleteGradeCategory(
+                $charon->category_id,
+                $charon->course
+            );
+
+            return $result;
+        }
+
+        return false;
     }
 
     /**
@@ -192,7 +203,14 @@ class CharonRepository
      */
     public function update($oldCharon, $newCharon)
     {
-        $modifiableFields = ['name', 'project_folder', 'tester_extra', 'system_extra', 'tester_type_code', 'grouping_id'];
+        $modifiableFields = ['name', 'project_folder', 'tester_extra', 'system_extra', 'tester_type_code',
+            'grouping_id', 'allow_submission'];
+
+        if (array_key_exists('allow_submission', $newCharon) and $newCharon['allow_submission'] == 'true') {
+            $newCharon['allow_submission'] = 1;
+        } else {
+            $newCharon['allow_submission'] = 0;
+        }
 
         $charon = $this->saveCharon($oldCharon, $newCharon, $modifiableFields);
 
@@ -265,7 +283,7 @@ class CharonRepository
                 },
             ])
                 ->where('charon_id', $charon->id)
-                ->get(['id', 'charon_id', 'grade_item_id', 'grade_type_code', 'name']);
+                ->get(['id', 'charon_id', 'grade_item_id', 'grade_type_code', 'name', 'persistent']);
             $charon->deadlines = Deadline::with([
                 'group' => function ($query) {
                     $query->select(['id', 'name']);
@@ -323,7 +341,7 @@ class CharonRepository
         $nullable_fields = ['docker_test_root', 'docker_content_root', 'system_extra', 'tester_extra'];
 
         foreach ($nullable_fields as $key) {
-            if (array_has($modifiableFields, $key)) {
+            if (Arr::has($modifiableFields, $key)) {
                 if (isset($updated[$key]) && $updated[$key] != '') {
                     $charon[$key] = $updated[$key];
                 } else {
@@ -333,7 +351,7 @@ class CharonRepository
         }
 
         foreach ($modifiableFields as $key) {
-            if (!array_has($nullable_fields, $key)) {
+            if (!Arr::has($nullable_fields, $key)) {
                 if (isset($updated[$key])) {
                     $charon[$key] = $updated[$key];
                 } else {
